// src/plugins/generated-shop-types.ts
export type Maybe<T> = T
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string | number
    String: string
    Boolean: boolean
    Int: number
    Float: number
    /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
    DateTime: any
    /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
    JSON: any
    /** The `Upload` scalar type represents a file upload. */
    Upload: any
}

export type ActiveOrderResult = NoActiveOrderError | Order

export type AddPaymentToOrderResult =
    | IneligiblePaymentMethodError
    | NoActiveOrderError
    | Order
    | OrderPaymentStateError
    | OrderStateTransitionError
    | PaymentDeclinedError
    | PaymentFailedError

export type Address = Node & {
    __typename?: 'Address'
    city?: Maybe<Scalars['String']>
    company?: Maybe<Scalars['String']>
    country: Country
    createdAt: Scalars['DateTime']
    customFields?: Maybe<Scalars['JSON']>
    defaultBillingAddress?: Maybe<Scalars['Boolean']>
    defaultShippingAddress?: Maybe<Scalars['Boolean']>
    fullName?: Maybe<Scalars['String']>
    id: Scalars['ID']
    phoneNumber?: Maybe<Scalars['String']>
    postalCode?: Maybe<Scalars['String']>
    province?: Maybe<Scalars['String']>
    streetLine1: Scalars['String']
    streetLine2?: Maybe<Scalars['String']>
    updatedAt: Scalars['DateTime']
}
// ...
export type AuthenticationMethod = Node & {
    __typename?: 'AuthenticationMethod'
    createdAt: Scalars['DateTime']
    id: Scalars['ID']
    strategy: Scalars['String']
    updatedAt: Scalars['DateTime']
}
// ...
export type Country = Node & {
    __typename?: 'Country'
    code: Scalars['String']
    createdAt: Scalars['DateTime']
    enabled: Scalars['Boolean']
    id: Scalars['ID']
    languageCode: LanguageCode
    name: Scalars['String']
    translations: Array<CountryTranslation>
    updatedAt: Scalars['DateTime']
}
// ...
export type Customer = Node & {
    __typename?: 'Customer'
    addresses?: Maybe<Array<Address>>
    createdAt: Scalars['DateTime']
    customFields?: Maybe<Scalars['JSON']>
    emailAddress: Scalars['String']
    firstName: Scalars['String']
    id: Scalars['ID']
    lastName: Scalars['String']
    orders: OrderList
    phoneNumber?: Maybe<Scalars['String']>
    title?: Maybe<Scalars['String']>
    updatedAt: Scalars['DateTime']
    user?: Maybe<User>
}
// ...
export enum ErrorCode {
    ALREADY_LOGGED_IN_ERROR = 'ALREADY_LOGGED_IN_ERROR',
    COUPON_CODE_EXPIRED_ERROR = 'COUPON_CODE_EXPIRED_ERROR',
    COUPON_CODE_INVALID_ERROR = 'COUPON_CODE_INVALID_ERROR',
    COUPON_CODE_LIMIT_ERROR = 'COUPON_CODE_LIMIT_ERROR',
    EMAIL_ADDRESS_CONFLICT_ERROR = 'EMAIL_ADDRESS_CONFLICT_ERROR',
    IDENTIFIER_CHANGE_TOKEN_EXPIRED_ERROR = 'IDENTIFIER_CHANGE_TOKEN_EXPIRED_ERROR',
    IDENTIFIER_CHANGE_TOKEN_INVALID_ERROR = 'IDENTIFIER_CHANGE_TOKEN_INVALID_ERROR',
    INELIGIBLE_PAYMENT_METHOD_ERROR = 'INELIGIBLE_PAYMENT_METHOD_ERROR',
    INELIGIBLE_SHIPPING_METHOD_ERROR = 'INELIGIBLE_SHIPPING_METHOD_ERROR',
    INSUFFICIENT_STOCK_ERROR = 'INSUFFICIENT_STOCK_ERROR',
    INVALID_CREDENTIALS_ERROR = 'INVALID_CREDENTIALS_ERROR',
    MISSING_PASSWORD_ERROR = 'MISSING_PASSWORD_ERROR',
    NATIVE_AUTH_STRATEGY_ERROR = 'NATIVE_AUTH_STRATEGY_ERROR',
    NEGATIVE_QUANTITY_ERROR = 'NEGATIVE_QUANTITY_ERROR',
    NOT_VERIFIED_ERROR = 'NOT_VERIFIED_ERROR',
    NO_ACTIVE_ORDER_ERROR = 'NO_ACTIVE_ORDER_ERROR',
    ORDER_LIMIT_ERROR = 'ORDER_LIMIT_ERROR',
    ORDER_MODIFICATION_ERROR = 'ORDER_MODIFICATION_ERROR',
    ORDER_PAYMENT_STATE_ERROR = 'ORDER_PAYMENT_STATE_ERROR',
    ORDER_STATE_TRANSITION_ERROR = 'ORDER_STATE_TRANSITION_ERROR',
    PASSWORD_ALREADY_SET_ERROR = 'PASSWORD_ALREADY_SET_ERROR',
    PASSWORD_RESET_TOKEN_EXPIRED_ERROR = 'PASSWORD_RESET_TOKEN_EXPIRED_ERROR',
    PASSWORD_RESET_TOKEN_INVALID_ERROR = 'PASSWORD_RESET_TOKEN_INVALID_ERROR',
    PAYMENT_DECLINED_ERROR = 'PAYMENT_DECLINED_ERROR',
    PAYMENT_FAILED_ERROR = 'PAYMENT_FAILED_ERROR',
    PHONE_NUMBER_INVALID_ERROR = 'PHONE_NUMBER_INVALID_ERROR',
    SMS_OTP_EXPIRED_ERROR = 'SMS_OTP_EXPIRED_ERROR',
    SMS_OTP_INVALID_ERROR = 'SMS_OTP_INVALID_ERROR',
    UNKNOWN_ERROR = 'UNKNOWN_ERROR',
    VERIFICATION_TOKEN_EXPIRED_ERROR = 'VERIFICATION_TOKEN_EXPIRED_ERROR',
    VERIFICATION_TOKEN_INVALID_ERROR = 'VERIFICATION_TOKEN_INVALID_ERROR',
}

export type ErrorResult = {
    errorCode: ErrorCode
    message: Scalars['String']
}
// ...
export type Node = {
    id: Scalars['ID']
}
// ...
export type Order = Node & {
    __typename?: 'Order'
    /** An order is active as long as the payment process has not been completed */
    active: Scalars['Boolean']
    billingAddress?: Maybe<OrderAddress>
    /** A unique code for the Order */
    code: Scalars['String']
    /** An array of all coupon codes applied to the Order */
    couponCodes: Array<Scalars['String']>
    createdAt: Scalars['DateTime']
    currencyCode: CurrencyCode
    customFields?: Maybe<Scalars['JSON']>
    customer?: Maybe<Customer>
    discounts: Array<Discount>
    fulfillments?: Maybe<Array<Fulfillment>>
    history: HistoryEntryList
    id: Scalars['ID']
    lines: Array<OrderLine>
    /**
     * The date & time that the Order was placed, i.e. the Customer
     * completed the checkout and the Order is no longer "active"
     */
    orderPlacedAt?: Maybe<Scalars['DateTime']>
    payments?: Maybe<Array<Payment>>
    /** Promotions applied to the order. Only gets populated after the payment process has completed. */
    promotions: Array<Promotion>
    shipping: Scalars['Int']
    shippingAddress?: Maybe<OrderAddress>
    shippingLines: Array<ShippingLine>
    shippingWithTax: Scalars['Int']
    state: Scalars['String']
    /**
     * The subTotal is the total of all OrderLines in the Order. This figure also includes any Order-level
     * discounts which have been prorated (proportionally distributed) amongst the OrderItems.
     * To get a total of all OrderLines which does not account for prorated discounts, use the
     * sum of `OrderLine.discountedLinePrice` values.
     */
    subTotal: Scalars['Int']
    /** Same as subTotal, but inclusive of tax */
    subTotalWithTax: Scalars['Int']
    /**
     * Surcharges are arbitrary modifications to the Order total which are neither
     * ProductVariants nor discounts resulting from applied Promotions. For example,
     * one-off discounts based on customer interaction, or surcharges based on payment
     * methods.
     */
    surcharges: Array<Surcharge>
    /** A summary of the taxes being applied to this Order */
    taxSummary: Array<OrderTaxSummary>
    /** Equal to subTotal plus shipping */
    total: Scalars['Int']
    totalQuantity: Scalars['Int']
    /** The final payable amount. Equal to subTotalWithTax plus shippingWithTax */
    totalWithTax: Scalars['Int']
    updatedAt: Scalars['DateTime']
}
// ...
export type OrderList = PaginatedList & {
    __typename?: 'OrderList'
    items: Array<Order>
    totalItems: Scalars['Int']
}
// ...
export type PaginatedList = {
    items: Array<Node>
    totalItems: Scalars['Int']
}
// ...
export type PhoneNumberInvalidError = ErrorResult & {
    __typename?: 'PhoneNumberInvalidError'
    errorCode: ErrorCode
    message: Scalars['String']
}
// ...
export type RequestCustomerSmsOtpResult = PhoneNumberInvalidError | Success

export type RequestPasswordResetResult = NativeAuthStrategyError | Success

export type RequestUpdateCustomerEmailAddressResult =
    | EmailAddressConflictError
    | InvalidCredentialsError
    | NativeAuthStrategyError
    | Success

export type ResetPasswordResult =
    | CurrentUser
    | NativeAuthStrategyError
    | PasswordResetTokenExpiredError
    | PasswordResetTokenInvalidError

export type Role = Node & {
    __typename?: 'Role'
    channels: Array<Channel>
    code: Scalars['String']
    createdAt: Scalars['DateTime']
    description: Scalars['String']
    id: Scalars['ID']
    permissions: Array<Permission>
    updatedAt: Scalars['DateTime']
}
// ...
export type SmsOtpExpiredError = ErrorResult & {
    __typename?: 'SmsOtpExpiredError'
    errorCode: ErrorCode
    message: Scalars['String']
}

export type SmsOtpInvalidError = ErrorResult & {
    __typename?: 'SmsOtpInvalidError'
    errorCode: ErrorCode
    message: Scalars['String']
}
// ...
export type Success = {
    __typename?: 'Success'
    success: Scalars['Boolean']
}
// ...
export type User = Node & {
    __typename?: 'User'
    authenticationMethods: Array<AuthenticationMethod>
    createdAt: Scalars['DateTime']
    customFields?: Maybe<Scalars['JSON']>
    id: Scalars['ID']
    identifier: Scalars['String']
    lastLogin?: Maybe<Scalars['DateTime']>
    roles: Array<Role>
    updatedAt: Scalars['DateTime']
    verified: Scalars['Boolean']
}
// ...
export type VerifyCustomerSmsOtpResult = Customer | SmsOtpExpiredError | SmsOtpInvalidError